Reorganized the code.
