+++ Indspire.Soaring.Engagement/Controllers/HomeController.cs
@@ -1,10 +1,9 @@
-﻿using Indspire.Soaring.Engagement.Data;
-using Indspire.Soaring.Engagement.Models;
-using Microsoft.AspNetCore.Mvc;
-using System.Diagnostics;
-
-namespace Indspire.Soaring.Engagement.Controllers
+﻿namespace Indspire.Soaring.Engagement.Controllers
 {
+    using System.Diagnostics;
+    using Indspire.Soaring.Engagement.Models;
+    using Microsoft.AspNetCore.Mvc;
+
     public class HomeController : Controller
     {
         public IActionResult Index()
@@ -17,7 +16,10 @@ namespace Indspire.Soaring.Engagement.Controllers
 
         public IActionResult Error()
         {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+            return View(new ErrorViewModel
+            {
+                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+            });
         }
     }
 }
