+++ Indspire.Soaring.Engagement/Extensions/EmailSenderExtensions.cs
@@ -1,12 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text.Encodings.Web;
-using System.Threading.Tasks;
-using Indspire.Soaring.Engagement.Services;
-
 namespace Indspire.Soaring.Engagement.Services
 {
+    using System.Text.Encodings.Web;
+    using System.Threading.Tasks;
+
     public static class EmailSenderExtensions
     {
         public static Task SendEmailConfirmationAsync(this IEmailSender emailSender, string email, string link)
