+++ Indspire.Soaring.Engagement/Extensions/UrlHelperExtensions.cs
@@ -1,11 +1,7 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Indspire.Soaring.Engagement.Controllers;
-
 namespace Microsoft.AspNetCore.Mvc
 {
+    using Indspire.Soaring.Engagement.Controllers;
+
     public static class UrlHelperExtensions
     {
         public static string EmailConfirmationLink(this IUrlHelper urlHelper, string userId, string code, string scheme)
