+++ Indspire.Soaring.Engagement/Extensions/UserExtensions.cs
@@ -1,12 +1,8 @@
-﻿using Indspire.Soaring.Engagement.Database;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Indspire.Soaring.Engagement.Utils;
-
-namespace Indspire.Soaring.Engagement.Extensions
+﻿namespace Indspire.Soaring.Engagement.Extensions
 {
+    using Indspire.Soaring.Engagement.Database;
+    using Indspire.Soaring.Engagement.Utils;
+
     public static class UserExtensions
     {
         public static string GetQRCodeAsBase64(this User user)
