+++ Indspire.Soaring.Engagement/ViewModels/JsonBaseViewModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Indspire.Soaring.Engagement.ViewModels
+﻿namespace Indspire.Soaring.Engagement.ViewModels
 {
     public class JsonBaseViewModel
     {
