+++ Indspire.Soaring.Engagement/ViewModels/LogActionJsonViewModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Indspire.Soaring.Engagement.ViewModels
+﻿namespace Indspire.Soaring.Engagement.ViewModels
 {
     public class LogActionJsonViewModel: JsonBaseViewModel
     {
@@ -16,12 +11,5 @@ namespace Indspire.Soaring.Engagement.ViewModels
         }
         public LogActionJsonViewModel() { }
         public LogActionResponseData ResponseData = new LogActionResponseData();
-    }
-
-    public class LogActionResponseData
-    {
-        public int PointsAwarded { get; set; }
-        public int PointsBalance { get; set; }
-        public string UserNumber { get; set; }
-    }
+    }   
 }
