+++ Indspire.Soaring.Engagement/ViewModels/LogActionResponseData.cs
@@ -1,0 +1,9 @@
+﻿namespace Indspire.Soaring.Engagement.ViewModels
+{
+    public class LogActionResponseData
+    {
+        public int PointsAwarded { get; set; }
+        public int PointsBalance { get; set; }
+        public string UserNumber { get; set; }
+    }
+}
