+++ Indspire.Soaring.Engagement/ViewModels/LogRedemptionJsonViewModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Indspire.Soaring.Engagement.ViewModels
+﻿namespace Indspire.Soaring.Engagement.ViewModels
 {
     public class LogRedemptionJsonViewModel : JsonBaseViewModel
     {
@@ -11,14 +6,7 @@ namespace Indspire.Soaring.Engagement.ViewModels
         {
             ResponseData = new LogRedemptionResponseData();
         }
-        public LogRedemptionResponseData ResponseData = new LogRedemptionResponseData();
-    }
 
-    public class LogRedemptionResponseData
-    {
-        public bool Success { get; set; }
-        public int PointsShort { get; set; }
-        public int PointsBalance { get; set; }
-        public string UserNumber { get; set; }
+        public LogRedemptionResponseData ResponseData = new LogRedemptionResponseData();
     }
 }
