+++ Indspire.Soaring.Engagement/ViewModels/LogRedemptionResponseData.cs
@@ -1,0 +1,10 @@
+﻿namespace Indspire.Soaring.Engagement.ViewModels
+{
+    public class LogRedemptionResponseData
+    {
+        public bool Success { get; set; }
+        public int PointsShort { get; set; }
+        public int PointsBalance { get; set; }
+        public string UserNumber { get; set; }
+    }
+}
